Move ProjectMenu
